refactor(goutil): split goutil.go monolith into concern-focused files
